+++ packages/core/src/components/index.ts
@@ -1,3 +1,4 @@
+export { default as Accordion } from './Accordion';
 export { default as Avatar } from './Avatar';
 export { default as BreadCrumbs } from './BreadCrumbs';
 export { default as Button } from './Button';
@@ -31,3 +32,4 @@ export { default as Toast } from './Toast';
 export { default as ToastContainer } from './ToastContainer';
 export { addToast } from './ToastContainer/ToastStore';
 export { default as Toggle } from './Toggle';
+