+++ packages/theme/src/core/accordion/index.ts
@@ -1,9 +1,0 @@
-import colors from "../colors";
-import { AccordionTheme } from "./types";
-
-const accordion: AccordionTheme = {
-    backgroundColor: colors.white,
-    separatorColor: colors.grey[100],
-}
-
-export default accordion;