+++ packages/theme/src/core/types.ts
@@ -1,5 +1,4 @@
 export * from './avatar/types';
-export * from './accordion/types';
 export * from './breadCrumbs/types';
 export * from './button/types';
 export * from './card/types';